check before acting on new folders for MKDIR function

+++ zfs-tools-functions.sh
@@ -351,12 +351,14 @@ MKDIR () {
     # Not all versions of GNU mkdir use the same characters around the directory names.
     # This seems to be fairly universal.
     mkdir --parents --verbose $folder | $AWK '{print $4}' | $SED 's/^.//' | $SED 's/.$//' > /tmp/new_folders_$$
-    new_folders=`cat /tmp/new_folders_$$`
-    for new_folder in $new_folders; do
-        chmod 2770 $new_folder
-        chgrp ozmt $new_folder
-    done    
-    rm -f /tmp/new_folders_$$
+    if [ -f /tmp/new_folders_$$ ]; then
+        new_folders=`cat /tmp/new_folders_$$`
+        for new_folder in $new_folders; do
+            chmod 2770 $new_folder
+            chgrp ozmt $new_folder
+        done    
+        rm -f /tmp/new_folders_$$
+    fi
 
 }
